Extract parts of `bootstrap::core::builder` into a `cli_paths` module
